Rename util files
